requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
blis==0.7.11
boto3==1.34.132
botocore==1.34.132
cachetools==5.3.3
catalogue==2.0.10
Cerberus==1.3.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
coverage==7.5.1
cryptography==42.0.5
cymem==2.0.8
defusedxml==0.8.0rc2
Django==5.0.4
django-storages==1.14.3
djangorestframework==3.15.0
factory-boy==3.3.0
Faker==25.0.1
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
kolo==2.24.0
langcodes==3.4.0
language_data==1.2.0
lightgbm==4.3.0
marisa-trie==1.1.0
MarkupSafe==2.1.5
msgpack==1.0.8
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pdfminer==20191125
pdfminer.six==20231228
pillow==10.2.0
pluggy==1.4.0
preshed==3.0.9
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pyparsing==3.1.2
pytest==8.1.1
pytest-cov==5.0.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
regex==2024.4.28
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.2
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
thinc==8.2.3
threadpoolctl==3.5.0
toolz==0.12.1
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.9.0
tzdata==2024.1
ulid-py==1.1.0
uritemplate==4.1.1
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
xgboost==2.0.3